Code review activity: Add random-number.py
This code demonstrates generally good naming, but there is an off by one error and repetitive comments.

Students probably do not know the behavior of random.random(), so I want them to take the initiative to look up the documentation. The hangman module does not exist, but its use is simple enought that it should be understandable. Students should get used to reading code that works with dependencies they are not familiar with.